backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
dnspython==2.4.2
cryptography==41.0.7
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1